feat: start working with sidebars
Initially, we will have product-specific sidebars at src/sidebars. Probably got to make an "in-app" mode to the whole project later, in which we get product-specific docsv2 stuff from integration-APIs if we are running under a deployment. So to make that a deployment won't have manuals on irrelevant things in it.